feat(screen): navigationServices was created and screen details repository was created.
